+++ packages/iov-bns/src/types.ts
@@ -31,40 +31,6 @@ export interface ChainAddressPair {
   readonly address: Address;
 }
 
-// blockchain NFT
-
-export interface BnsBlockchainNft {
-  readonly id: string;
-  readonly owner: Address;
-  /**
-   * The registered chain information
-   *
-   * Fields as defined in https://github.com/iov-one/bns-spec/blob/master/docs/data/ObjectDefinitions.rst#chain
-   */
-  readonly chain: {
-    readonly chainId: ChainId;
-    readonly name: string;
-    readonly enabled: boolean;
-    readonly production: boolean;
-    readonly networkId: string | undefined;
-    readonly mainTickerId: TokenTicker | undefined;
-  };
-  readonly codecName: string;
-  readonly codecConfig: string;
-}
-
-export interface BnsBlockchainsByChainIdQuery {
-  readonly chainId: ChainId;
-}
-
-export type BnsBlockchainsQuery = BnsBlockchainsByChainIdQuery;
-
-export function isBnsBlockchainsByChainIdQuery(
-  query: BnsBlockchainsQuery,
-): query is BnsBlockchainsByChainIdQuery {
-  return typeof query.chainId !== "undefined";
-}
-
 // username NFT
 
 export interface BnsUsernameNft {