+++ packages/iov-bns/types/index.d.ts
@@ -2,4 +2,4 @@ export { bnsCodec } from "./bnscodec";
 export { bnsConnector } from "./bnsconnector";
 export { BnsConnection } from "./bnsconnection";
 export { bnsSwapQueryTag } from "./tags";
-export { ChainAddressPair, Participant, BnsBlockchainNft, BnsBlockchainsByChainIdQuery, BnsBlockchainsQuery, BnsUsernamesByChainAndAddressQuery, BnsUsernamesByOwnerAddressQuery, BnsUsernamesByUsernameQuery, BnsUsernamesQuery, BnsUsernameNft, BnsTx, AddAddressToUsernameTx, CreateMultisignatureTx, RegisterUsernameTx, RemoveAddressFromUsernameTx, UpdateMultisignatureTx, } from "./types";
+export { ChainAddressPair, Participant, BnsUsernamesByChainAndAddressQuery, BnsUsernamesByOwnerAddressQuery, BnsUsernamesByUsernameQuery, BnsUsernamesQuery, BnsUsernameNft, BnsTx, AddAddressToUsernameTx, CreateMultisignatureTx, RegisterUsernameTx, RemoveAddressFromUsernameTx, UpdateMultisignatureTx, } from "./types";